ref(node): Move dns into `internals`
